fix merchant invoice show page without breaking tests

+++ db/migrate/20230421013752_create_bulk_discounts.rb
@@ -3,7 +3,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
     create_table :bulk_discounts do |t|
       t.references :merchant, foreign_key: true
       t.integer :threshold
-      t.integer :discount
+      t.float :discount
 
       t.timestamps
     end
